requirements.txt
asgiref==3.8.1
black==24.8.0
click==8.1.7
colorama==0.4.6
Django==5.1.1
django-extensions==3.2.3
django-filter==24.3
django-otp==1.5.4
django-unfold==0.39.0
djangorestframework==3.15.2
djangorestframework-jsonapi==7.0.2
inflection==0.5.1
mypy-extensions==1.0.0
packaging==24.1
pathspec==0.12.1
pillow==10.4.0
platformdirs==4.3.6
psycopg2==2.9.9
python-decouple==3.8
pytz==2024.2
sqlparse==0.5.1
tzdata==2024.2
